Version 3.1/requirements.txt
streamlit
pandas
beautifulsoup4
requests
langchain
langchain_core
langchain-community
langchain-openai
langgraph
openai
tabulate
openpyxl
reportlab
lxml
python-dotenv
urllib3
crawl4ai
html2text
nest_asyncio
playwright
scikit-learn
datasketch
sentence-transformers
faiss-cpu
